src/104B/test_cdf_104B.py
import unittest
from unittest.mock import patch

from cdf_104B import CodeforcesTask104BSolution


class TestCDF104B(unittest.TestCase):
    def test_104B_acceptance_1(self):
        mock_input = ['2', '1 1']
        expected = '2'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask104BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_104B_acceptance_2(self):
        mock_input = ['2', '2 2']
        expected = '5'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask104BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_104B_acceptance_3(self):
        mock_input = ['1', '10']
        expected = '10'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask104BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)


if __name__ == "__main__":
    unittest.main()
